User, Role and Auth refactor.
These packages followed a different architecture, I refactored them to
follow the new architecture.

	renamed:    demo/src/main/java/com/project/demo/entity/auth/ApplicationConfiguration.java -> demo/src/main/java/com/project/demo/ApplicationConfiguration.java
	renamed:    demo/src/main/java/com/project/demo/entity/auth/SecurityConfiguration.java -> demo/src/main/java/com/project/demo/SecurityConfiguration.java
	renamed:    demo/src/main/java/com/project/demo/entity/user/LoginResponse.java -> demo/src/main/java/com/project/demo/entity/LoginResponse.java
	renamed:    demo/src/main/java/com/project/demo/entity/rol/Role.java -> demo/src/main/java/com/project/demo/entity/Role.java
	renamed:    demo/src/main/java/com/project/demo/entity/rol/RoleEnum.java -> demo/src/main/java/com/project/demo/entity/RoleEnum.java
	renamed:    demo/src/main/java/com/project/demo/entity/user/User.java -> demo/src/main/java/com/project/demo/entity/User.java
	renamed:    demo/src/main/java/com/project/demo/entity/auth/AuthenticationService.java -> demo/src/main/java/com/project/demo/logic/AuthenticationService.java
	renamed:    demo/src/main/java/com/project/demo/entity/auth/JwtAuthenticationFilter.java -> demo/src/main/java/com/project/demo/logic/JwtAuthenticationFilter.java
	renamed:    demo/src/main/java/com/project/demo/entity/auth/JwtService.java -> demo/src/main/java/com/project/demo/logic/JwtService.java
	renamed:    demo/src/main/java/com/project/demo/entity/user/UserRepository.java -> demo/src/main/java/com/project/demo/logic/UserRepository.java
	renamed:    demo/src/main/java/com/project/demo/entity/rol/AdminSeeder.java -> demo/src/main/java/com/project/demo/logic/seeder/AdminSeeder.java
	renamed:    demo/src/main/java/com/project/demo/entity/rol/RoleSeeder.java -> demo/src/main/java/com/project/demo/logic/seeder/RoleSeeder.java
	modified:   demo/src/main/java/com/project/demo/repository/CurrencyRepository.java
	renamed:    demo/src/main/java/com/project/demo/entity/rol/RoleRepository.java -> demo/src/main/java/com/project/demo/repository/RoleRepository.java
	renamed:    demo/src/main/java/com/project/demo/rest/admin/AdminController.java -> demo/src/main/java/com/project/demo/rest/AdminController.java
	renamed:    demo/src/main/java/com/project/demo/rest/auth/AuthRestController.java -> demo/src/main/java/com/project/demo/rest/AuthRestController.java
	renamed:    demo/src/main/java/com/project/demo/rest/country/CountryRestController.java -> demo/src/main/java/com/project/demo/rest/CountryRestController.java
	renamed:    demo/src/main/java/com/project/demo/rest/user/UserRestController.java -> demo/src/main/java/com/project/demo/rest/UserRestController.java
	modified:   demo/src/test/java/com/project/demo/Config/TestConfig.java
	modified:   demo/src/test/java/com/project/demo/repository/UserRepositoryTests.java
