+++ demo/src/main/java/com/project/demo/entity/RoleEnum.java
@@ -1,4 +1,4 @@
-package com.project.demo.entity.rol;
+package com.project.demo.entity;
 
 public enum RoleEnum {
     USER,
