+++ demo/src/main/java/com/project/demo/repository/CurrencyRepository.java
@@ -1,8 +1,6 @@
 package com.project.demo.repository;
 
 import com.project.demo.entity.Currency;
-import com.project.demo.entity.rol.Role;
-import com.project.demo.entity.rol.RoleEnum;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
