+++ demo/src/test/java/com/project/demo/Config/TestConfig.java
@@ -1,8 +1,8 @@
 package com.project.demo.Config;
 
 
-import com.project.demo.entity.auth.AuthenticationService;
-import com.project.demo.entity.user.UserRepository;
+import com.project.demo.logic.AuthenticationService;
+import com.project.demo.logic.UserRepository;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
